requirements.travis.txt
pyyaml
astropy
pysftp